+++ CHANGELOG.md
@@ -444,3 +444,4 @@
 * Define `P_T` joint schema exclusively for `Prism` type
 * Define `Prism` type and its `preview` method
 * [WIP] Redefine `Lens` type so it's has the same scheme as `Prism`
+* Define `Optics` type as generalization of `Lens` and `Prism`